% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millefy-bamImport.R
\name{calcBamNormFactors}
\alias{calcBamNormFactors}
\title{Calculate normalization factors for visualization of BAM files}
\usage{
calcBamNormFactors(bam_files, scaleFactor = 10^6)
}
\arguments{
\item{bam_files}{A character vector containing paths to BAM files}

\item{scaleFactor}{A numeric vector of normalization factors (1/(total mapped reads)*scaleFactor)}
}
\value{
bamNormFactors numeric vector
}
\description{
Calculate normalization factors for visualization of BAM files
}
\examples{
# Gene annotation track (For faster performance, try to use \\code{dt_gtf} paramter)
bam_files = Sys.glob(file.path(system.file("extdata/bam", package="millefy"), "*.bam"))
normFactors = calcBamNormFactors(bam_files)
}
