% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millefy-genemodel.R
\name{gtfToDtExon}
\alias{gtfToDtExon}
\title{Create a data.table object that contains informaiton on exonic regions}
\usage{
gtfToDtExon(path_gtf)
}
\arguments{
\item{path_gtf}{Path to a GTF file}
}
\value{
dtExon data.table object
}
\description{
Create a data.table object that contains informaiton on exonic regions
}
\examples{
# Gene annotation track (For faster performance, try to use \\code{dt_gtf} paramter)
path_gtf = "annotation/gencode.gtf"
dt_gtf_exon <- gtfToDtExon(path_gtf) 
geneTrack1 <- list(path_gtf = path_gtf, dt_gtf = dt_gtf_exon, label = "GENCODE")
}
