% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millefy-lastplot.R
\name{millefy_last_plot}
\alias{millefy_last_plot}
\title{Retrieve the last Millefy plot to be modified or created.}
\usage{
millefy_last_plot()
}
\description{
Retrieve the last Millefy plot to be modified or created.
}
\keyword{internal}
