% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millefy-lastplot.R
\name{set_millefy_last_plot}
\alias{set_millefy_last_plot}
\title{Set the last Millefy plot to be fetched by millefy_last_plot()}
\usage{
set_millefy_last_plot(value)
}
\description{
Set the last Millefy plot to be fetched by millefy_last_plot()
}
\seealso{
[millefy_last_plot()]
}
\keyword{internal}
